add TimeTree::BaseClient
