+++ lib/timetree/calendar_app/client.rb
@@ -6,8 +6,7 @@ require 'jwt'
 module TimeTree
   module CalendarApp
     # TimeTree API CalendarApp client.
-    class Client
-      API_HOST = 'https://timetreeapis.com'
+    class Client < BaseClient
       # @return [Integer]
       attr_reader :installation_id
       # @return [String]
@@ -16,12 +15,6 @@ module TimeTree
       attr_reader :private_key
       # @return [String]
       attr_reader :token
-      # @return [Integer]
-      attr_reader :ratelimit_limit
-      # @return [Integer]
-      attr_reader :ratelimit_remaining
-      # @return [Time]
-      attr_reader :ratelimit_reset_at
 
       # @param installation_id [Integer] CalendarApp's installation id
       # @param application_id [String] CalendarApp id
@@ -111,16 +104,7 @@ module TimeTree
         activity
       end
 
-      def update_ratelimit(res)
-        limit = res.headers['x-ratelimit-limit']
-        remaining = res.headers['x-ratelimit-remaining']
-        reset = res.headers['x-ratelimit-reset']
-        @ratelimit_limit = limit.to_i if limit
-        @ratelimit_remaining = remaining.to_i if remaining
-        @ratelimit_reset_at = Time.at reset.to_i if reset
-      end
-
-    private
+      private
 
       attr_reader :http_cmd, :access_token
 
@@ -156,29 +140,6 @@ module TimeTree
           iss: application_id
         }
         JWT.encode(payload, private_key, 'RS256')
-      end
-
-      def check_event_id(value)
-        check_required_property(value, 'event_id')
-      end
-
-      def check_required_property(value, name)
-        err = Error.new "#{name} is required."
-        raise err if value.nil?
-        raise err if value.to_s.empty?
-
-        true
-      end
-
-      def to_model(data, included: nil)
-        TimeTree::BaseModel.to_model data, client: self, included: included
-      end
-
-      def relationships_params(relationships, default)
-        params = {}
-        relationships ||= default
-        params[:include] = relationships.join ',' if relationships.is_a? Array
-        params
       end
     end
   end
